requirements.txt
Flask==3.0.0
PyYAML==6.0.1
pika==1.3.2
redis==5.0.1
mysql-connector-python==8.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0

